requirements.txt
attrs==20.3.0
Automat==20.2.0
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
constantly==15.1.0
Flask==1.1.2
Flask-GoogleCharts==0.1.0
Flask-MySQL==1.5.2
Flask-MySQLdb==0.2.0
gunicorn==20.0.4
hyperlink==21.0.0
idna==2.10
incremental==17.5.0
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mysql==0.0.2
mysqlclient==1.4.6
npm==0.1.1
optional-django==0.1.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
PyMySQL==1.0.2
python-dateutil==2.8.1
python-decouple==3.4
pytz==2020.5
regex==2020.11.13
requests==2.25.1
six==1.15.0
txaio==20.12.1
tzlocal==2.1
ujson==4.0.1
urllib3==1.26.2
Werkzeug==1.0.1
zope.interface==5.2.0
